loop based on json not char files
Now it is looped by json objects and not char files - it is more simple with operating with filtered json. Json data from specific orgin (e.g. rcam) are filtered and only some of data are saved to filtered json based on seed, because there is a huge amount of rcam multipliers - algorithm is more faster.